+++ packages/cloud-cognitive/src/components/ActionBar/ActionBar.js
@@ -181,6 +181,7 @@ export let ActionBar = React.forwardRef(
       // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [maxVisible, itemArray]);
 
+    /* istanbul ignore next */ // not sure how to fake window resize
     const handleResize = () => {
       // width is the space available for all action bar items horizontally
       // the action bar items are squares so the height should be one item wide
@@ -188,8 +189,10 @@ export let ActionBar = React.forwardRef(
       checkFullyVisibleItems();
     };
 
+    /* istanbul ignore next */ // not sure how to fake window resize
     const handleActionBarItemsResize = () => {
       // when the hidden sizing items change size
+      /* istanbul ignore next */ // not sure how to fake window resize
       checkFullyVisibleItems();
     };