+++ packages/cloud-cognitive/src/components/index.js
@@ -6,7 +6,6 @@
 //
 
 export { AboutModal } from './AboutModal';
-export { ActionBarItem } from './ActionBar';
 export { APIKeyModal } from './APIKeyModal';
 export { ComboButton, ComboButtonItem } from './ComboButton';
 export { ContextHeader } from './ContextHeader';